requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
tensorflow==2.15.0
pandas==2.1.4
matplotlib==3.8.2
reportlab==4.0.7
PyPDF2==3.0.1
python-magic-bin==0.4.14
exifread==3.0.0
pywavelets==1.5.0
joblib==1.3.2
